% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent.R
\name{calc_tool}
\alias{calc_tool}
\title{Example Calculator Tool}
\usage{
calc_tool(expr)
}
\arguments{
\item{expr}{A string containing a mathematical expression}
}
\value{
The result of the calculation
}
\description{
A simple example tool that can perform basic calculations
}
