% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{default_chat_model}
\alias{default_chat_model}
\title{Get default chat model}
\usage{
default_chat_model()
}
\value{
Character string with the default model name
}
\description{
Returns the default model to use for chat interactions.
Currently defaults to "llama3.2:1b".
}
