% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kb_tool.R
\name{get_embedding}
\alias{get_embedding}
\title{Get Embedding}
\usage{
get_embedding(text)
}
\arguments{
\item{text}{Text to get embedding for}
}
\value{
Numeric vector containing the embedding
}
\description{
Gets an embedding vector for a text using the Ollama API.
}
\keyword{internal}
