% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kb_tool.R
\name{kb_tool}
\alias{kb_tool}
\title{Knowledge Base Tool}
\usage{
kb_tool(
  dir = "~/knowledgebase",
  n_context = 3,
  update = FALSE,
  show_chunks = FALSE
)
}
\arguments{
\item{dir}{Directory containing markdown or text files}

\item{n_context}{Number of similar texts to use as context (default: 3)}

\item{update}{If TRUE, force update of vector store}

\item{show_chunks}{Whether to display the chunks used for the answer}
}
\value{
A function that can be used as a tool with create_agent()
}
\description{
A tool for querying a knowledge base using vector embeddings.
}
