% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ollama.R
\name{ollama_client}
\alias{ollama_client}
\title{Create an Ollama client}
\usage{
ollama_client(model, system_prompt = NULL, format = NULL)
}
\arguments{
\item{model}{Character string specifying the model to use}

\item{system_prompt}{Character string with the system prompt}

\item{format}{Optional list specifying the JSON schema for structured output}
}
\value{
A list containing the client's methods
}
\description{
Create an Ollama client
}
